Move CMPTally out of omnicore.h
No functional change.
