+++ src/Makefile.omnicore.include
@@ -18,6 +18,7 @@ OMNICORE_H = \
   omnicore/script.h \
   omnicore/sp.h \
   omnicore/sto.h \
+  omnicore/tally.h \
   omnicore/tx.h \
   omnicore/utils.h \
   omnicore/version.h \
@@ -42,6 +43,7 @@ OMNICORE_CPP = \
   omnicore/script.cpp \
   omnicore/sp.cpp \
   omnicore/sto.cpp \
+  omnicore/tally.cpp \
   omnicore/tx.cpp \
   omnicore/utils.cpp \
   omnicore/version.cpp \
