+++ src/omnicore/omnicore.h
@@ -1,8 +1,3 @@
-// global Master Protocol header file
-// globals consts (DEFINEs for now) should be here
-//
-// for now (?) class declarations go here -- work in progress; probably will get pulled out into a separate file, note: some declarations are still in the .cpp file
-
 #ifndef OMNICORE_OMNICORE_H
 #define OMNICORE_OMNICORE_H
 
@@ -12,6 +7,7 @@ class CTransaction;
 
 #include "omnicore/log.h"
 #include "omnicore/persistence.h"
+#include "omnicore/tally.h"
 
 #include "sync.h"
 #include "uint256.h"
@@ -166,197 +162,6 @@ extern bool autoCommit;
 
 //! Global lock for state objects
 extern CCriticalSection cs_tally;
-
-//! Balance record types
-enum TallyType { BALANCE = 0, SELLOFFER_RESERVE = 1, ACCEPT_RESERVE = 2, PENDING = 3, METADEX_RESERVE = 4, TALLY_TYPE_COUNT };
-
-/** Balance records of a single entity.
- */
-class CMPTally
-{
-private:
-    typedef struct {
-        int64_t balance[TALLY_TYPE_COUNT];
-    } BalanceRecord;
-
-    //! Map of balance records
-    typedef std::map<uint32_t, BalanceRecord> TokenMap;
-    //! Balance records for different tokens
-    TokenMap mp_token;
-    //! Internal iterator pointing to a balance record
-    TokenMap::iterator my_it;
-
-public:
-    /** Resets the internal iterator. */
-    uint32_t init()
-    {
-        uint32_t ret = 0;
-        my_it = mp_token.begin();
-        if (my_it != mp_token.end()) {
-            ret = my_it->first;
-        }
-        return ret;
-    }
-
-    /** Advances the internal iterator. */
-    uint32_t next()
-    {
-        uint32_t ret = 0;
-        if (my_it != mp_token.end()) {
-            ret = my_it->first;
-            ++my_it;
-        }
-        return ret;
-    }
-
-    /** Updates the number of tokens for the given tally type. */
-    bool updateMoney(uint32_t propertyId, int64_t amount, TallyType ttype)
-    {
-        if (TALLY_TYPE_COUNT <= ttype || amount == 0) {
-            return false;
-        }
-        bool bRet = false;
-        int64_t now64 = mp_token[propertyId].balance[ttype];
-
-        if (PENDING != ttype && (now64 + amount) < 0) {
-            // NOTE:
-            // This check also serves as overflow protection!
-        } else {
-            now64 += amount;
-            mp_token[propertyId].balance[ttype] = now64;
-
-            bRet = true;
-        }
-
-        return bRet;
-    }
-
-    /** Creates an empty tally. */
-    CMPTally()
-    {
-        my_it = mp_token.begin();
-    }
-
-    /** Returns the number of tokens for the given tally type. */
-    int64_t getMoney(uint32_t propertyId, TallyType ttype) const
-    {
-        if (TALLY_TYPE_COUNT <= ttype) {
-            return 0;
-        }
-        int64_t money = 0;
-        TokenMap::const_iterator it = mp_token.find(propertyId);
-
-        if (it != mp_token.end()) {
-            const BalanceRecord& record = it->second;
-            money = record.balance[ttype];
-        }
-
-        return money;
-    }
-
-    /** Returns the number of available tokens. */
-    int64_t getMoneyAvailable(uint32_t propertyId) const
-    {
-        TokenMap::const_iterator it = mp_token.find(propertyId);
-
-        if (it != mp_token.end()) {
-            const BalanceRecord& record = it->second;
-            if (record.balance[PENDING] < 0) {
-                return record.balance[BALANCE] + record.balance[PENDING];
-            } else {
-                return record.balance[BALANCE];
-            }
-        }
-
-        return 0;
-    }
-
-    /** Returns the number of reserved tokens. */
-    int64_t getMoneyReserved(uint32_t propertyId) const
-    {
-        int64_t money = 0;
-        TokenMap::const_iterator it = mp_token.find(propertyId);
-
-        if (it != mp_token.end()) {
-            const BalanceRecord& record = it->second;
-            money += record.balance[SELLOFFER_RESERVE];
-            money += record.balance[ACCEPT_RESERVE];
-            money += record.balance[METADEX_RESERVE];
-        }
-
-        return money;
-    }
-
-    /** Compares the tally with another tally and returns true, if they are equal. */
-    bool operator==(const CMPTally& rhs) const
-    {
-        if (mp_token.size() != rhs.mp_token.size()) {
-            return false;
-        }
-        TokenMap::const_iterator pc1 = mp_token.begin();
-        TokenMap::const_iterator pc2 = rhs.mp_token.begin();
-
-        for (unsigned int i = 0; i < mp_token.size(); ++i) {
-            if (pc1->first != pc2->first) {
-                return false;
-            }
-            const BalanceRecord& record1 = pc1->second;
-            const BalanceRecord& record2 = pc2->second;
-
-            for (int ttype = 0; ttype < TALLY_TYPE_COUNT; ++ttype) {
-                if (record1.balance[ttype] != record2.balance[ttype]) {
-                    return false;
-                }
-            }
-            ++pc1;
-            ++pc2;
-        }
-
-        assert(pc1 == mp_token.end());
-        assert(pc2 == rhs.mp_token.end());
-
-        return true;
-    }
-
-    /** Compares the tally with another tally and returns true, if they are not equal. */
-    bool operator!=(const CMPTally& rhs) const
-    {
-        return !operator==(rhs);
-    }
-
-    /** Prints a balance record to the console. */
-    int64_t print(uint32_t propertyId = OMNI_PROPERTY_MSC, bool bDivisible = true) const
-    {
-        int64_t balance = 0;
-        int64_t selloffer_reserve = 0;
-        int64_t accept_reserve = 0;
-        int64_t pending = 0;
-        int64_t metadex_reserve = 0;
-
-        TokenMap::const_iterator it = mp_token.find(propertyId);
-
-        if (it != mp_token.end()) {
-            const BalanceRecord& record = it->second;
-            balance = record.balance[BALANCE];
-            selloffer_reserve = record.balance[SELLOFFER_RESERVE];
-            accept_reserve = record.balance[ACCEPT_RESERVE];
-            pending = record.balance[PENDING];
-            metadex_reserve = record.balance[METADEX_RESERVE];
-        }
-
-        if (bDivisible) {
-            PrintToConsole("%22s [ SO_RESERVE= %22s, ACCEPT_RESERVE= %22s, METADEX_RESERVE= %22s ] %22s\n",
-                    FormatDivisibleMP(balance, true), FormatDivisibleMP(selloffer_reserve, true),
-                    FormatDivisibleMP(accept_reserve, true), FormatDivisibleMP(metadex_reserve, true),
-                    FormatDivisibleMP(pending, true));
-        } else {
-            PrintToConsole("%14d [ SO_RESERVE= %14d, ACCEPT_RESERVE= %14d, METADEX_RESERVE= %14d ] %14d\n",
-                    balance, selloffer_reserve, accept_reserve, metadex_reserve, pending);
-        }
-
-        return (balance + selloffer_reserve + accept_reserve + metadex_reserve);
-    }
-};
 
 /** LevelDB based storage for STO recipients.
  */
