+++ src/omnicore/walletcache.cpp
@@ -3,6 +3,7 @@
 
 #include "omnicore/log.h"
 #include "omnicore/omnicore.h"
+#include "omnicore/tally.h"
 
 #include "init.h"
 #include "sync.h"
