+++ src/qt/sendmpdialog.cpp
@@ -16,6 +16,7 @@
 #include "omnicore/parse_string.h"
 #include "omnicore/pending.h"
 #include "omnicore/sp.h"
+#include "omnicore/tally.h"
 
 #include "amount.h"
 #include "base58.h"
